Outsource watchlistbuttons to own component
